Combine file models into one
