+++ resumable/file.py
@@ -1,6 +1,9 @@
 import os
+import uuid
 from threading import Lock
 from collections import namedtuple
+
+from resumable.util import CallbackDispatcher
 
 
 FileChunk = namedtuple('FileChunk', ['index', 'size', 'read'])
@@ -26,18 +29,22 @@ def build_chunks(read_bytes, file_size, chunk_size):
     return chunks
 
 
-class LazyLoadChunkableFile(object):
+class ResumableFile(object):
 
     def __init__(self, path, chunk_size):
+
         self.path = str(path)
+        self.unique_identifier = uuid.uuid4()
         self.chunk_size = int(chunk_size)
+        self.size = os.path.getsize(path)
 
-        self.size = os.path.getsize(self.path)
-
-        self._fp = open(self.path, 'rb')
+        self._fp = open(path, 'rb')
         self._fp_lock = Lock()
 
         self.chunks = build_chunks(self._read_bytes, self.size, chunk_size)
+        self.chunk_done = {chunk: False for chunk in self.chunks}
+
+        self.completed = CallbackDispatcher()
 
     def close(self):
         self._fp.close()
@@ -46,3 +53,12 @@ class LazyLoadChunkableFile(object):
         with self._fp_lock:
             self._fp.seek(start)
             return self._fp.read(num_bytes)
+
+    @property
+    def is_completed(self):
+        return all(self.chunk_done.values())
+
+    def handle_chunk_completion(self):
+        if self.is_completed:
+            self.completed.trigger()
+            self.close()
